+++ api/exchange.go
@@ -3,6 +3,8 @@ package api
 import (
 	"fmt"
 
+	"github.com/stellar/go/build"
+	"github.com/stellar/go/clients/horizon"
 	"github.com/stellar/kelp/model"
 )
 
@@ -14,7 +16,7 @@ type ExchangeAPIKey struct {
 
 // Account allows you to access key account functions
 type Account interface {
-	GetAccountBalances(assetList []model.Asset) (map[model.Asset]model.Number, error)
+	GetAccountBalances(assetList []interface{}) (map[interface{}]model.Number, error)
 }
 
 // Ticker encapsulates all the data for a given Trading Pair
@@ -73,17 +75,22 @@ type Constrainable interface {
 	GetOrderConstraints(pair *model.TradingPair) *model.OrderConstraints
 }
 
+// OrderbookFetcher extracts out the method that should go into ExchangeShim for now
+type OrderbookFetcher interface {
+	GetOrderBook(pair *model.TradingPair, maxCount int32) (*model.OrderBook, error)
+}
+
 // TradeAPI is the interface we use as a generic API for trading on any crypto exchange
 type TradeAPI interface {
 	GetAssetConverter() *model.AssetConverter
 
 	Constrainable
 
-	GetOrderBook(pair *model.TradingPair, maxCount int32) (*model.OrderBook, error)
+	OrderbookFetcher
 
 	GetTrades(pair *model.TradingPair, maybeCursor interface{}) (*TradesResult, error)
 
-	TradeFetcher
+	FillTrackable
 
 	GetOpenOrders(pairs []*model.TradingPair) (map[model.TradingPair][]model.OpenOrder, error)
 
@@ -190,4 +197,21 @@ type Exchange interface {
 	TradeAPI
 	DepositAPI
 	WithdrawAPI
+}
+
+// Balance repesents various aspects of an asset's balance
+type Balance struct {
+	Balance float64
+	Trust   float64
+	Reserve float64
+}
+
+// ExchangeShim is the interface we use as a generic API for all crypto exchanges
+type ExchangeShim interface {
+	SubmitOps(ops []build.TransactionMutator, asyncCallback func(hash string, e error)) error
+	GetBalanceHack(asset horizon.Asset) (*Balance, error)
+	LoadOffersHack() ([]horizon.Offer, error)
+	Constrainable
+	OrderbookFetcher
+	FillTrackable
 }