+++ cmd/exchanges.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/stellar/kelp/plugins"
+	"github.com/stellar/kelp/support/sdk"
 
 	"github.com/spf13/cobra"
 )
@@ -15,11 +16,13 @@ var exchanagesCmd = &cobra.Command{
 
 func init() {
 	exchanagesCmd.Run = func(ccmd *cobra.Command, args []string) {
-		fmt.Printf("  Exchange\t\tSupports Trading\tDescription\n")
+		// call sdk.GetExchangeList() here so we pre-load exchanges before displaying the table
+		sdk.GetExchangeList()
+		fmt.Printf("  Exchange\t\t\tTested\t\tTrading\t\tDescription\n")
 		fmt.Printf("  --------------------------------------------------------------------------------\n")
 		exchanges := plugins.Exchanges()
 		for _, name := range sortedExchangeKeys(exchanges) {
-			fmt.Printf("  %-14s\t%v\t\t\t%s\n", name, exchanges[name].TradeEnabled, exchanges[name].Description)
+			fmt.Printf("  %-24s\t%v\t\t%v\t\t%s\n", name, exchanges[name].Tested, exchanges[name].TradeEnabled, exchanges[name].Description)
 		}
 	}
 }