+++ plugins/factory.go
@@ -9,12 +9,14 @@ import (
 	"github.com/stellar/kelp/api"
 	"github.com/stellar/kelp/model"
 	"github.com/stellar/kelp/plugins/p2pb2b"
+	"github.com/stellar/kelp/support/sdk"
 	"github.com/stellar/kelp/support/utils"
 )
 
 // strategyFactoryData is a data container that has all the information needed to make a strategy
 type strategyFactoryData struct {
 	sdex            *SDEX
+	ieif            *IEIF
 	tradingPair     *model.TradingPair
 	assetBase       *horizon.Asset
 	assetQuote      *horizon.Asset
@@ -33,7 +35,7 @@ type StrategyContainer struct {
 
 // strategies is a map of all the strategies available
 var strategies = map[string]StrategyContainer{
-	"buysell": StrategyContainer{
+	"buysell": {
 		SortOrder:   1,
 		Description: "Creates buy and sell offers based on a reference price with a pre-specified liquidity depth",
 		NeedsConfig: true,
@@ -43,14 +45,14 @@ var strategies = map[string]StrategyContainer{
 			err := config.Read(strategyFactoryData.stratConfigPath, &cfg)
 			utils.CheckConfigError(cfg, err, strategyFactoryData.stratConfigPath)
 			utils.LogConfig(cfg)
-			s, e := makeBuySellStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg)
+			s, e := makeBuySellStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.ieif, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg)
 			if e != nil {
 				return nil, fmt.Errorf("makeFn failed: %s", e)
 			}
 			return s, nil
 		},
 	},
-	"mirror": StrategyContainer{
+	"mirror": {
 		SortOrder:   4,
 		Description: "Mirrors an orderbook from another exchange by placing the same orders on Stellar",
 		NeedsConfig: true,
@@ -60,14 +62,14 @@ var strategies = map[string]StrategyContainer{
 			err := config.Read(strategyFactoryData.stratConfigPath, &cfg)
 			utils.CheckConfigError(cfg, err, strategyFactoryData.stratConfigPath)
 			utils.LogConfig(cfg)
-			s, e := makeMirrorStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg, strategyFactoryData.simMode)
+			s, e := makeMirrorStrategy(strategyFactoryData.sdex, strategyFactoryData.ieif, strategyFactoryData.tradingPair, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg, strategyFactoryData.simMode)
 			if e != nil {
 				return nil, fmt.Errorf("makeFn failed: %s", e)
 			}
 			return s, nil
 		},
 	},
-	"sell": StrategyContainer{
+	"sell": {
 		SortOrder:   0,
 		Description: "Creates sell offers based on a reference price with a pre-specified liquidity depth",
 		NeedsConfig: true,
@@ -77,14 +79,14 @@ var strategies = map[string]StrategyContainer{
 			err := config.Read(strategyFactoryData.stratConfigPath, &cfg)
 			utils.CheckConfigError(cfg, err, strategyFactoryData.stratConfigPath)
 			utils.LogConfig(cfg)
-			s, e := makeSellStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg)
+			s, e := makeSellStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.ieif, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg)
 			if e != nil {
 				return nil, fmt.Errorf("makeFn failed: %s", e)
 			}
 			return s, nil
 		},
 	},
-	"balanced": StrategyContainer{
+	"balanced": {
 		SortOrder:   3,
 		Description: "Dynamically prices two tokens based on their relative demand",
 		NeedsConfig: true,
@@ -94,10 +96,10 @@ var strategies = map[string]StrategyContainer{
 			err := config.Read(strategyFactoryData.stratConfigPath, &cfg)
 			utils.CheckConfigError(cfg, err, strategyFactoryData.stratConfigPath)
 			utils.LogConfig(cfg)
-			return makeBalancedStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg), nil
+			return makeBalancedStrategy(strategyFactoryData.sdex, strategyFactoryData.tradingPair, strategyFactoryData.ieif, strategyFactoryData.assetBase, strategyFactoryData.assetQuote, &cfg), nil
 		},
 	},
-	"delete": StrategyContainer{
+	"delete": {
 		SortOrder:   2,
 		Description: "Deletes all orders for the configured orderbook",
 		NeedsConfig: false,
@@ -111,6 +113,7 @@ var strategies = map[string]StrategyContainer{
 // MakeStrategy makes a strategy
 func MakeStrategy(
 	sdex *SDEX,
+	ieif *IEIF,
 	tradingPair *model.TradingPair,
 	assetBase *horizon.Asset,
 	assetQuote *horizon.Asset,
@@ -119,12 +122,14 @@ func MakeStrategy(
 	simMode bool,
 ) (api.Strategy, error) {
 	log.Printf("Making strategy: %s\n", strategy)
-	if strat, ok := strategies[strategy]; ok {
-		if strat.NeedsConfig && stratConfigPath == "" {
+	if s, ok := strategies[strategy]; ok {
+		if s.NeedsConfig && stratConfigPath == "" {
 			return nil, fmt.Errorf("the '%s' strategy needs a config file", strategy)
 		}
-		s, e := strat.makeFn(strategyFactoryData{
+
+		s, e := s.makeFn(strategyFactoryData{
 			sdex:            sdex,
+			ieif:            ieif,
 			tradingPair:     tradingPair,
 			assetBase:       assetBase,
 			assetQuote:      assetQuote,
@@ -153,91 +158,77 @@ type exchangeFactoryData struct {
 
 // ExchangeContainer contains the exchange factory method along with some metadata
 type ExchangeContainer struct {
-	SortOrder    uint8
+	SortOrder    uint16
 	Description  string
 	TradeEnabled bool
+	Tested       bool
 	makeFn       func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error)
 }
 
 // exchanges is a map of all the exchange integrations available
-var exchanges = map[string]ExchangeContainer{
-	"kraken": ExchangeContainer{
-		SortOrder:    0,
-		Description:  "Kraken is a popular centralized cryptocurrency exchange (https://www.kraken.com/)",
-		TradeEnabled: true,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return makeKrakenExchange(exchangeFactoryData.apiKeys, exchangeFactoryData.simMode)
+var exchanges *map[string]ExchangeContainer
+
+// getExchanges returns a map of all the exchange integrations available
+func getExchanges() map[string]ExchangeContainer {
+	if exchanges == nil {
+		loadExchanges()
+	}
+	return *exchanges
+}
+
+func loadExchanges() {
+	// marked as tested if key exists in this map (regardless of bool value)
+	testedCcxtExchanges := map[string]bool{
+		"binance": true,
+	}
+
+	exchanges = &map[string]ExchangeContainer{
+		"kraken": {
+			SortOrder:    0,
+			Description:  "Kraken is a popular centralized cryptocurrency exchange",
+			TradeEnabled: true,
+			Tested:       true,
+			makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
+				return makeKrakenExchange(exchangeFactoryData.apiKeys, exchangeFactoryData.simMode)
+			},
 		},
-	},
-	"ccxt-kraken": ExchangeContainer{
-		SortOrder:    1,
-		Description:  "Kraken is a popular centralized cryptocurrency exchange (via ccxt-rest)",
-		TradeEnabled: false,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return makeCcxtExchange(
-				"http://localhost:3000",
-				"kraken",
-				nil,
-				exchangeFactoryData.apiKeys,
-				exchangeFactoryData.simMode,
-			)
+		"p2pb2b": {
+			SortOrder:    5,
+			Description:  "P2PB2B is an Estonian cryptocurrency exchange (https://p2pb2b.io/)",
+			TradeEnabled: true,
+			makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
+				return p2pb2b.MakeP2PB2BExchange(exchangeFactoryData.apiKeys, exchangeFactoryData.simMode)
+			},
 		},
-	},
-	"ccxt-binance": ExchangeContainer{
-		SortOrder:    2,
-		Description:  "Binance is a popular centralized cryptocurrency exchange (via ccxt-rest)",
-		TradeEnabled: true,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return makeCcxtExchange(
-				"http://localhost:3000",
-				"binance",
-				nil,
-				exchangeFactoryData.apiKeys,
-				exchangeFactoryData.simMode,
-			)
-		},
-	},
-	"ccxt-poloniex": ExchangeContainer{
-		SortOrder:    3,
-		Description:  "Poloniex is a popular centralized cryptocurrency exchange (via ccxt-rest)",
-		TradeEnabled: false,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return makeCcxtExchange(
-				"http://localhost:3000",
-				"poloniex",
-				nil,
-				exchangeFactoryData.apiKeys,
-				exchangeFactoryData.simMode,
-			)
-		},
-	},
-	"ccxt-bittrex": ExchangeContainer{
-		SortOrder:    4,
-		Description:  "Bittrex is a popular centralized cryptocurrency exchange (via ccxt-rest)",
-		TradeEnabled: false,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return makeCcxtExchange(
-				"http://localhost:3000",
-				"bittrex",
-				nil,
-				exchangeFactoryData.apiKeys,
-				exchangeFactoryData.simMode,
-			)
-		},
-	},
-	"p2pb2b": ExchangeContainer{
-		SortOrder:    5,
-		Description:  "P2PB2B is an Estonian cryptocurrency exchange (https://p2pb2b.io/)",
-		TradeEnabled: true,
-		makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
-			return p2pb2b.MakeP2PB2BExchange(exchangeFactoryData.apiKeys, exchangeFactoryData.simMode)
-		},
-	},
+	}
+
+	// add all CCXT exchanges
+	sortOrderOffset := len(*exchanges)
+	for i, exchangeName := range sdk.GetExchangeList() {
+		key := fmt.Sprintf("ccxt-%s", exchangeName)
+		_, tested := testedCcxtExchanges[exchangeName]
+		boundExchangeName := exchangeName
+
+		(*exchanges)[key] = ExchangeContainer{
+			SortOrder:    uint16(i + sortOrderOffset),
+			Description:  exchangeName + " is automatically added via ccxt-rest",
+			TradeEnabled: true,
+			Tested:       tested,
+			makeFn: func(exchangeFactoryData exchangeFactoryData) (api.Exchange, error) {
+				return makeCcxtExchange(
+					boundExchangeName,
+					nil,
+					exchangeFactoryData.apiKeys,
+					exchangeFactoryData.simMode,
+				)
+			},
+		}
+	}
 }
 
 // MakeExchange is a factory method to make an exchange based on a given type
 func MakeExchange(exchangeType string, simMode bool) (api.Exchange, error) {
-	if exchange, ok := exchanges[exchangeType]; ok {
+	if exchange, ok := getExchanges()[exchangeType]; ok {
 		exchangeAPIKey := api.ExchangeAPIKey{Key: "", Secret: ""}
 		x, e := exchange.makeFn(exchangeFactoryData{
 			simMode: simMode,
@@ -254,7 +245,7 @@ func MakeExchange(exchangeType string, simMode bool) (api.Exchange, error) {
 
 // MakeTradingExchange is a factory method to make an exchange based on a given type
 func MakeTradingExchange(exchangeType string, apiKeys []api.ExchangeAPIKey, simMode bool) (api.Exchange, error) {
-	if exchange, ok := exchanges[exchangeType]; ok {
+	if exchange, ok := getExchanges()[exchangeType]; ok {
 		if !exchange.TradeEnabled {
 			return nil, fmt.Errorf("trading is not enabled on this exchange: %s", exchangeType)
 		}
@@ -278,5 +269,5 @@ func MakeTradingExchange(exchangeType string, apiKeys []api.ExchangeAPIKey, simM
 
 // Exchanges returns the list of exchanges
 func Exchanges() map[string]ExchangeContainer {
-	return exchanges
+	return getExchanges()
 }