+++ support/sdk/ccxt.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"net/http"
 	"reflect"
+	"strconv"
 	"strings"
 
 	"github.com/mitchellh/mapstructure"
@@ -14,10 +15,12 @@ import (
 	"github.com/stellar/kelp/support/networking"
 )
 
+// ccxtBaseURL should not have suffix of '/'
+const ccxtBaseURL = "http://localhost:3000"
+
 // Ccxt Rest SDK (https://github.com/franz-see/ccxt-rest, https://github.com/ccxt/ccxt/)
 type Ccxt struct {
 	httpClient   *http.Client
-	ccxtBaseURL  string
 	exchangeName string
 	instanceName string
 	markets      map[string]CcxtMarket
@@ -36,6 +39,9 @@ type CcxtMarket struct {
 		Price struct {
 			Min float64 `json:"min"`
 		} `json:"price"`
+		Cost struct {
+			Min float64 `json:"min"`
+		} `json:"cost"`
 	} `json:"limits"`
 	Precision struct {
 		Amount int8 `json:"amount"`
@@ -46,7 +52,7 @@ type CcxtMarket struct {
 const pathExchanges = "/exchanges"
 
 // MakeInitializedCcxtExchange constructs an instance of Ccxt that is bound to a specific exchange instance on the CCXT REST server
-func MakeInitializedCcxtExchange(ccxtBaseURL string, exchangeName string, apiKey api.ExchangeAPIKey) (*Ccxt, error) {
+func MakeInitializedCcxtExchange(exchangeName string, apiKey api.ExchangeAPIKey) (*Ccxt, error) {
 	if strings.HasSuffix(ccxtBaseURL, "/") {
 		return nil, fmt.Errorf("invalid format for ccxtBaseURL: %s", ccxtBaseURL)
 	}
@@ -57,12 +63,11 @@ func MakeInitializedCcxtExchange(ccxtBaseURL string, exchangeName string, apiKey
 	}
 	c := &Ccxt{
 		httpClient:   http.DefaultClient,
-		ccxtBaseURL:  ccxtBaseURL,
 		exchangeName: exchangeName,
 		instanceName: instanceName,
 	}
 
-	e = c.init(apiKey)
+	e = c.initialize(apiKey)
 	if e != nil {
 		return nil, fmt.Errorf("error when initializing Ccxt exchange: %s", e)
 	}
@@ -70,29 +75,50 @@ func MakeInitializedCcxtExchange(ccxtBaseURL string, exchangeName string, apiKey
 	return c, nil
 }
 
-func (c *Ccxt) init(apiKey api.ExchangeAPIKey) error {
-	// get exchange list
-	var exchangeList []string
-	e := networking.JSONRequest(c.httpClient, "GET", c.ccxtBaseURL+pathExchanges, "", map[string]string{}, &exchangeList, "error")
-	if e != nil {
-		return fmt.Errorf("error getting list of supported exchanges by CCXT: %s", e)
-	}
+// exchangeList contains a list of supported exchanges
+var exchangeList *[]string
 
+// GetExchangeList gets a list of all supported exchanges
+func GetExchangeList() []string {
+	if exchangeList == nil {
+		loadExchangeList()
+	}
+	return *exchangeList
+}
+
+func loadExchangeList() {
+	var output []string
+	e := networking.JSONRequest(http.DefaultClient, "GET", ccxtBaseURL+pathExchanges, "", map[string]string{}, &output, "error")
+	if e != nil {
+		eMsg1 := strings.Contains(e.Error(), "could not execute http request")
+		eMsg2 := strings.Contains(e.Error(), "http://localhost:3000/exchanges: dial tcp")
+		eMsg3 := strings.Contains(e.Error(), "connection refused")
+		if eMsg1 && eMsg2 && eMsg3 {
+			log.Printf("ccxt-rest is not running on port 3000 so we cannot include those exchanges")
+		} else {
+			panic(fmt.Errorf("error getting list of supported exchanges by CCXT: %s", e))
+		}
+	}
+	exchangeList = &output
+}
+
+func (c *Ccxt) initialize(apiKey api.ExchangeAPIKey) error {
 	// validate that exchange name is in the exchange list
 	exchangeListed := false
-	for _, name := range exchangeList {
+	el := GetExchangeList()
+	for _, name := range el {
 		if name == c.exchangeName {
 			exchangeListed = true
 			break
 		}
 	}
 	if !exchangeListed {
-		return fmt.Errorf("exchange name '%s' is not in the list of %d exchanges available: %v", c.exchangeName, len(exchangeList), exchangeList)
+		return fmt.Errorf("exchange name '%s' is not in the list of %d exchanges available: %v", c.exchangeName, len(el), el)
 	}
 
 	// list all the instances of the exchange
 	var instanceList []string
-	e = networking.JSONRequest(c.httpClient, "GET", c.ccxtBaseURL+pathExchanges+"/"+c.exchangeName, "", map[string]string{}, &instanceList, "error")
+	e := networking.JSONRequest(c.httpClient, "GET", ccxtBaseURL+pathExchanges+"/"+c.exchangeName, "", map[string]string{}, &instanceList, "error")
 	if e != nil {
 		return fmt.Errorf("error getting list of exchange instances for exchange '%s': %s", c.exchangeName, e)
 	}
@@ -110,7 +136,7 @@ func (c *Ccxt) init(apiKey api.ExchangeAPIKey) error {
 
 	// load markets to populate fields related to markets
 	var marketsResponse interface{}
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/loadMarkets"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/loadMarkets"
 	e = networking.JSONRequest(c.httpClient, "POST", url, "", map[string]string{}, &marketsResponse, "error")
 	if e != nil {
 		return fmt.Errorf("error loading markets for exchange instance (exchange=%s, instanceName=%s): %s", c.exchangeName, c.instanceName, e)
@@ -171,7 +197,7 @@ func (c *Ccxt) newInstance(apiKey api.ExchangeAPIKey) error {
 	}
 
 	var newInstance map[string]interface{}
-	e = networking.JSONRequest(c.httpClient, "POST", c.ccxtBaseURL+pathExchanges+"/"+c.exchangeName, string(data), map[string]string{}, &newInstance, "error")
+	e = networking.JSONRequest(c.httpClient, "POST", ccxtBaseURL+pathExchanges+"/"+c.exchangeName, string(data), map[string]string{}, &newInstance, "error")
 	if e != nil {
 		return fmt.Errorf("error in web request when creating new exchange instance for exchange '%s': %s", c.exchangeName, e)
 	}
@@ -185,7 +211,7 @@ func (c *Ccxt) newInstance(apiKey api.ExchangeAPIKey) error {
 // symbolExists returns an error if the symbol does not exist
 func (c *Ccxt) symbolExists(tradingPair string) error {
 	// get list of symbols available on exchange
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var exchangeOutput interface{}
 	e := networking.JSONRequest(c.httpClient, "GET", url, "", map[string]string{}, &exchangeOutput, "error")
@@ -231,7 +257,7 @@ func (c *Ccxt) FetchTicker(tradingPair string) (map[string]interface{}, error) {
 	}
 
 	// fetch ticker for symbol
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchTicker"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchTicker"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -271,7 +297,7 @@ func (c *Ccxt) FetchOrderBook(tradingPair string, limit *int) (map[string][]Ccxt
 	}
 
 	// fetch orderbook for symbol
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchOrderBook"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchOrderBook"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -311,6 +337,10 @@ type CcxtTrade struct {
 	Side      string  `json:"side"`
 	Symbol    string  `json:"symbol"`
 	Timestamp int64   `json:"timestamp"`
+	Fee       struct {
+		Cost     float64 `json:"cost"`
+		Currency string  `json:"currency"`
+	} `json:"fee"`
 }
 
 // FetchTrades calls the /fetchTrades endpoint on CCXT, trading pair is the CCXT version of the trading pair
@@ -328,7 +358,7 @@ func (c *Ccxt) FetchTrades(tradingPair string) ([]CcxtTrade, error) {
 	}
 
 	// fetch trades for symbol
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchTrades"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchTrades"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	output := []CcxtTrade{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -338,19 +368,30 @@ func (c *Ccxt) FetchTrades(tradingPair string) ([]CcxtTrade, error) {
 	return output, nil
 }
 
-func (c *Ccxt) FetchMyTrades(tradingPair string) ([]CcxtTrade, error) {
+// FetchMyTrades calls the /fetchMyTrades endpoint on CCXT, trading pair is the CCXT version of the trading pair
+func (c *Ccxt) FetchMyTrades(tradingPair string, limit int, maybeCursorStart interface{}) ([]CcxtTrade, error) {
 	e := c.symbolExists(tradingPair)
 	if e != nil {
 		return nil, fmt.Errorf("symbol does not exist: %s", e)
 	}
 
 	// marshal input data
-	data, e := json.Marshal(&[]string{tradingPair})
-	if e != nil {
-		return nil, fmt.Errorf("error marshaling input (tradingPair=%s) as an array for exchange '%s': %s", tradingPair, c.exchangeName, e)
+	var data []byte
+	if maybeCursorStart == nil {
+		data, e = json.Marshal(&[]string{tradingPair, strconv.Itoa(limit)})
+		if e != nil {
+			return nil, fmt.Errorf("error marshaling input (tradingPair=%s) as an array for exchange '%s': %s", tradingPair, c.exchangeName, e)
+		}
+	} else {
+		cursorString := fmt.Sprintf("%v", maybeCursorStart)
+		data, e = json.Marshal(&[]string{tradingPair, cursorString, strconv.Itoa(limit)})
+		if e != nil {
+			return nil, fmt.Errorf("error marshaling input (tradingPair=%s, maybeCursorStart=%v) as an array for exchange '%s': %s", tradingPair, maybeCursorStart, c.exchangeName, e)
+		}
 	}
+
 	// fetch trades for symbol
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchMyTrades"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchMyTrades"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	output := []CcxtTrade{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -369,7 +410,7 @@ type CcxtBalance struct {
 
 // FetchBalance calls the /fetchBalance endpoint on CCXT
 func (c *Ccxt) FetchBalance() (map[string]CcxtBalance, error) {
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchBalance"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchBalance"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e := networking.JSONRequest(c.httpClient, "POST", url, "", map[string]string{}, &output, "error")
@@ -435,7 +476,7 @@ func (c *Ccxt) FetchOpenOrders(tradingPairs []string) (map[string][]CcxtOpenOrde
 		return nil, fmt.Errorf("error marshaling input (tradingPairs=%v) for exchange '%s': %s", tradingPairs, c.exchangeName, e)
 	}
 
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchOpenOrders"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/fetchOpenOrders"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -491,7 +532,7 @@ func (c *Ccxt) CreateLimitOrder(tradingPair string, side string, amount float64,
 		return nil, fmt.Errorf("error marshaling input (%v) for exchange '%s': %s", inputData, c.exchangeName, e)
 	}
 
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/createOrder"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/createOrder"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")
@@ -530,7 +571,7 @@ func (c *Ccxt) CancelOrder(orderID string, tradingPair string) (*CcxtOpenOrder, 
 		return nil, fmt.Errorf("error marshaling input (%v) for exchange '%s': %s", inputData, c.exchangeName, e)
 	}
 
-	url := c.ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/cancelOrder"
+	url := ccxtBaseURL + pathExchanges + "/" + c.exchangeName + "/" + c.instanceName + "/cancelOrder"
 	// decode generic data (see "https://blog.golang.org/json-and-go#TOC_4.")
 	var output interface{}
 	e = networking.JSONRequest(c.httpClient, "POST", url, string(data), map[string]string{}, &output, "error")